requirements.txt
langchain==0.0.329
openai==0.28.1
discord.py==2.3.2
python-dotenv==1.00
pymupdf==1.23.6
tiktoken==0.5.1
logzero==1.7.0